feat(runtime): widen UINode/RichRun/KeyframeStop/KeyStyle/Binding color fields uint16->uint32 (888-ready); add 888 blend/lerp alongside; keep 565 math for TFT byte-identity

+++ packages/cuttlefish/src/emit/emitters/ui-emitter.ts
@@ -193,7 +193,7 @@ export function emitUIRuntime(ctx: EmitterContext): void {
       const access = spec.property === "background" ? "bg" : spec.property === "color" ? "fg" : "bg";
       const body = spec.cppExpr || `__ui_nodes[${spec.nodeIndex}].${access}`;
       ctx.sourceLines.push(
-        `uint16_t ${spec.fnName}(void) { return ${body}; }`,
+        `uint32_t ${spec.fnName}(void) { return ${body}; }`,
       );
     }
   }
